+++ app/models/beneficiaries/Beneficiary.scala
@@ -31,11 +31,30 @@ trait IncomeBeneficiary extends Beneficiary {
   val incomeDiscretionYesNo: Option[Boolean]
   val countryOfResidence: Option[String]
   val address: Option[Address]
+
+  override def hasRequiredData(migratingFromNonTaxableToTaxable: Boolean, trustType: Option[TypeOfTrust]): Boolean = {
+    if (migratingFromNonTaxableToTaxable) {
+      (incomeDiscretionYesNo, income) match {
+        case (None, None) => false
+        case _ => true
+      }
+    } else {
+      true
+    }
+  }
 }
 
 trait OrgBeneficiary extends IncomeBeneficiary {
   val name: String
   val utr: Option[String]
+
+  override def hasRequiredData(migratingFromNonTaxableToTaxable: Boolean, trustType: Option[TypeOfTrust]): Boolean = {
+    if (utr.isEmpty) {
+      super.hasRequiredData(migratingFromNonTaxableToTaxable, trustType)
+    } else {
+      true
+    }
+  }
 }
 
 trait BeneficiaryReads {
