+++ app/models/beneficiaries/CompanyBeneficiary.scala
@@ -16,7 +16,7 @@
 
 package models.beneficiaries
 
-import models.{Address, TypeOfTrust}
+import models.Address
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
@@ -29,19 +29,7 @@ case class CompanyBeneficiary(name: String,
                               incomeDiscretionYesNo: Option[Boolean],
                               countryOfResidence: Option[String] = None,
                               entityStart: LocalDate,
-                              provisional: Boolean) extends OrgBeneficiary {
-
-  override def hasRequiredData(migratingFromNonTaxableToTaxable: Boolean, trustType: Option[TypeOfTrust]): Boolean = {
-    if (migratingFromNonTaxableToTaxable && utr.isEmpty) {
-      (incomeDiscretionYesNo, income) match {
-        case (None, None) => false
-        case _ => true
-      }
-    } else {
-      true
-    }
-  }
-}
+                              provisional: Boolean) extends OrgBeneficiary
 
 object CompanyBeneficiary extends BeneficiaryReads {
